Trait pallet_robonomics_staking::pallet::Config

pub trait Config: Config {
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type OnBond: OnBondHandler<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type BondingDuration: Get<Self::BlockNumber>;
    type StakeReward: Get<Perbill>;
    type BonusReward: Get<Perbill>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

The staking balance.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type OnBond: OnBondHandler<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

New bonded value handler.

type BondingDuration: Get<Self::BlockNumber>

Some time that staked funds must remain bonded for.

type StakeReward: Get<Perbill>

Standard stake reward for 1 XRT each block. Note: Perbill has accuracy 10^9, that equal to Wn.

type BonusReward: Get<Perbill>

Bonus stake income for 1 XRT each block. Note: Perbill has accuracy 10^9, that equal to Wn.